feat: add flaky tests to the test results PR comment
The Failure object passed in the MessagePayload now contains 2 extra
fields: flags and flake.

the flags field are the flags associated with the upload of the failure.
the field is a list of strings

the flake field represents the flake on the Test this failure is on.
this field is a list of Flake objects.

A Flake object consists of 2 fields: symptoms (a list of enums that
represent what kind of flake this test is a part of) and is_new_flake
(signifies whether the flake was a previously know one, or if it was
just detected)

This commit also moves all the sample test result files to the samples
directory in the tests directory.

Signed-off-by: joseph-sentry <joseph.sawaya@sentry.io>